Fixed Bugs And Updated Formatting
Fixed bug whereby only one photo was deleted even if it had a copy.

Prevented creation of copies if the photo is/ has a copy.

Reintroduced the expand all button and removed the custom expand row button.

Removed the ability to expand to show the picture for copies of rows.

Centered the picture in the expanded row